add basic auth

+++ src/main/java/com/example/book_quiz/repository/SavedBookRepository.java
@@ -3,6 +3,6 @@ package com.example.book_quiz.repository;
 import com.example.book_quiz.entity.SavedBookEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SavedBookRepository extends CrudRepository<SavedBookEntity, Integer> {
+public interface SavedBookRepository extends CrudRepository<SavedBookEntity, Long> {
 
 }
